eval context refactoring
